Speed-up of statistics
git-svn-id: file:///home/roberto/Git/svn@6799 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: b8b23e10f4d1150f63fd9ee24fbb51152b6cc7bc

+++ trunk/xmipp/libraries/data/multidim_array.h
@@ -2834,7 +2834,7 @@ public:
         FOR_ALL_DIRECT_ELEMENTS_IN_MULTIDIMARRAY_ptr(*this,n,ptr)
         {
             T Tval=*ptr;
-            double val=static_cast< double >(Tval);
+            double val=Tval;
             avg += val;
             stddev += val * val;
